Create Phase2_Prepare
This is Phase 2 of the Google Data Analysis process.

This R code installs and loads various packages required for data analysis, then sets the working directory to where the CSV files are located. The code then reads the "cyclisticdata.csv" file, removes missing and duplicate values, creates new columns for ride duration, day of the week, time of day, and season, and saves the new dataset as "cleanedcyclisticdata3.csv" and "cleanedcyclisticdata3.rda". The code also displays information about the structure, summary statistics, and column names of the data at various stages of the cleaning process.