Phase2_MergingDatasets
This is the start of Phase 2 in the Google Data Analysis process.

This code merges multiple yearly datasets of Cyclistic bike share data and checks the validity of the combined data. The code installs and attaches several R packages such as tidyverse, janitor, data.validator, lubridate, data.table, dplyr, knitr, and readr. 

The code then sets the working directory to where the CSV files are located, gets a list of all CSV files in the directory that contain data from the years 2020, 2021, and 2022, and reads and combines them using dplyr::bind_rows(). 

The merged data is then viewed and summarized using summary() function. Finally, the merged data is saved as a CSV file named "cyclisticdata.csv".